rgw: prevent notify and error_repo from duplicating bucket sync
The marker_tracker generally prevents us from starting two sync
operations on the same bucket, but there are two cases where we start
bucket sync without having a marker:
* in response to async datalog notifications, and
* when retrying buckets in the error_repo

This commit augments RGWDataSyncShardMarkerTrack so that these two cases
can still register their sync operation, so that a later attempt knows
not to start again. If, after starting, we get a valid marker from the
log, we attach it to the sync in progress so that marker is written on
completion.

This should help reduce contention on the sync status locks, which will
in turn reduce the number of writes we make to the error_repo.

Signed-off-by: Casey Bodley <cbodley@redhat.com>