osd/PrimaryLogPG: send requests to primary on cache miss
If a client has {BALANCE,LOCALIZE}_READS and sends a request to a
replica, but the object isn't in the cache, send them back to the
primary.  Otherwise we might do something rash (like trigger a
promotion from a replica).

Fixes: http://tracker.ceph.com/issues/20919
Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit 741a8720996b74434a036b143209111ce5203cbc)