mon/PGMap: use new-style options
Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit 7a23097c34c35c1bf6ec09e86ed3acbb0807068c)

Conflicts:
	src/common/legacy_config_opts.h
	src/mon/PGMap.cc: in 729a089, PGMap::get_health() is removed
from master branch, but in luminous, this function is still around for
backward compatibility. but that function is still using the old-style
options. in this change, i choose to keep some of the old-style options
removed in the cherry-picked cleanup to appease PGMap::get_health(),
instead of migrating all of them to the new-style because of less
efforts.