osd: move peering events into primary work queue
For the moment we leave the old infratructure in place too; that will be
cleaned up next.

Two key differences:

- each pg event is processed in isolation, which means MOSDPGNotify etc
messages won't be batched.  This will mean more messages for small clusters
and no change for large clusters.
- the map processing for PGs will be done in completion by a single event.
There is now a wait in handle_osd_map to ensure that all PGs are processing
maps that overlap with the map cache.

Signed-off-by: Sage Weil <sage@redhat.com>