osd/PG: allow preemption of scrub due to client request
Drop the PG lock momentarily between scrub chunks.  When we retake the lock,
check for a peering change and scrub preemption.

If a client op (op or repop) arrives while we are scrubbing a range, set a
preempt flag.  In the scrub code, exit quickly if the atomic flag gets set.

Once we complete the scrub cycle, if the preempt flag was set, restart the
range.  Allow this to happen up to N times.  Once we hit the limit, block
client IO in the scrub range like we did before.

Signed-off-by: Sage Weil <sage@redhat.com>