osd: use final map for deleted pools for split pg_num checks
We keep PGs around after their pools are deleted.  Keep the final OSDMap
for the pool around in order to make the split checks' lives easier.  In
particular, _maybe_split_pgid() just needs the last map for the pool in
order to tell whether it needs to split the splitting pg_t's.

Signed-off-by: Sage Weil <sage@redhat.com>