msg/async: do not make Stack a fork watcher
Instead of doing a stop/start on fork, shut down entirely pre-fork.  This
is a cleaner solution for the cases where e.g. the parent does not need
to restart the msgr at all (this is usually the case!) but does need
to call psotfork_start in order to restart log thread (which is assumed
to be running when the parent's cct is destroyed).

Signed-off-by: Sage Weil <sage@redhat.com>