osd: fix advance_pg when pool dne during starting epoch
We may have already advanced beyond a pools deletion and need to advance
some more.  In that case, OSDMap::get_pg_num() will assert out if the
pool does not exit.  Skip the pg split and merge checks in this case just
like we do when the pool doesn't exist in the new map.

Signed-off-by: Sage Weil <sage@redhat.com>