mds: change libmon to .la archive
before this change, it is a plain .a archive. because MDSAuthCaps.cc is
included by both libmon.a and libmds.la, automake complains that

src/Makefile.am: error: object 'mds/MDSAuthCaps.$(OBJEXT)' created both
with libtool and without

so we need to consolidate them. as most of the convenience libraries are
are created using libtool. to be more consistent, i am chaning libmon.a
to libmon.la.

Signed-off-by: Kefu Chai <kchai@redhat.com>
Conflicts:
   src/Makefile-env.am
   src/mds/Makefile-server.am: this change is not cherry-picked from
master. as we switched from autotools to cmake since luminous.