osd: fix _process handling for pg vs slot race
We could see the slot with a different PG than we expected if the old
PG was removed and a new one was instantiated in its place.  We can't
just pick up the new PG pointer, however, since it isn't locked.

Fix by just bailing out.  The current event was intended for the original
PG, which no longer exists.

Signed-off-by: Sage Weil <sage@redhat.com>