osd/OSDMap.cc: remove pg_upmap/pg_upmap_items too if osd is gone
If an osd is gone or moved out from the specific crush rule,
we should cancel any pg_upmap/pg_upmap_items still bound to
that osd too.

The original code does not work for the above case because
get_parent_of_type() will fail first if that osd does not belong
to the crush_rule passed in.

Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>