common/ceph_mutex: ceph::{mutex,condition_variable,lock_guard}
If CEPH_DEBUG_MUTEX is defined, use the legacy Mutex class that implements
Ceph's lockdep and a bunch of other random debugging machinery.  Also
typedef ceph::condition_variable to std::condition_variable_any, which will
work with any Lockable mutex (including Mutex), and lock_guard, which
is defined as std::lock_guard<Mutex>.

If CEPH_DEBUG_MUTEX is not defined, then use the bare-bones C++ std::mutex
primitives... or as close as we can get to them.  Since we take a
arguments to the Mutex constructor, do the same for ceph::mutex, but (in
this compilation mode) ignore them and simply wrap the std::mutex.
Provide an operator std::mutex&() so that ceph::unique_lock can be
defined as std::unique_lock<std::mutex> and we can pass in a ceph::mutex
to the constructor.  (It's important that we use unique_lock<std::mutex>
so that we can use std::condition_variable, which is more efficient
than std::condition_variable_any.)

Signed-off-by: Sage Weil <sage@redhat.com>