rgw: batch and process bucket instances
Sort through and batch bucket instances so that multiple calls to
reading current bucket info and locking can be avoided. For the most
trivial case when the bucket is already deleted we exit early with all
the stale instances. When the bucket reshard is in progress we only
process the stale entries with status done, if the bucket is available
for locking then we lock down and mark the other instances as well.

Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>
(cherry picked from commit fb9c0497621fce34283094cfe260480f9122573f)

Resolves: rhbz#1653830