rgw: check for non-existent bucket in RGWGetACLs
Fix incorrect result when attempting to fetch an ACL for a
bucket that does not exist.

This change leads to a NoSuchBucket error result when an ACL is
requested from a non-existent bucket.  A NoSuchKey error result
iss already seen when the target was an object.

Fixes: https://tracker.ceph.com/issues/38116
Resolves: rhbz#1667142

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>
(cherry picked from commit 078981ca4ef1610ab89b60f4c281f521d6838a37)
Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>