msg/async: always queue_reset on connection stop
We are attaching a Session to incoming (v2) OSD connections
unconditionally (on authentication), so we need to make sure we call
ms_handle_reset in all cases... including when we mark_down() locally.

Simplify the V1 and V2 code by moving the queue_reset() call into stop(),
and simplify the external callers to rely on it (instead of calling
queue_reset themselves).

I can't think of any reason we *wouldn't* mind having ms_handle_reset
called, unless the dispatcher was disconnected before shutdown--but that
is not allowed and would cause a zillion other bad things.

Signed-off-by: Sage Weil <sage@redhat.com>