msg/async/AsyncMessenger: clear Connection::priv during reap
We frequently have ref cycles between Connection and various session
objects.  When it comes time to discard the connectino, make sure we
break the priv half of the cycle so that we don't leak memory.

Note that in the OSD case we normally break the cycle in the *other*
direction during ms_handle_reset in order to avoid races with
the backoff handling code that looks things up by session.  By the time
we get to reap we assume that ms_handle_reset has already done it's
existing thing and broken the cycle in the cases that matter.  This
patch adds a catch-all for anything that leaks through, for example due
to connections that start (v2) authentication but don't finish getting
set up and registered, or slip through the cracks in some other way.

We can't simply clear priv in the reap method because dispatch_queue
events may still be in flight (usually handle_reset).  Instead, queue
a final event that clears priv to ensure that it sequences *after* any
other events and is really the last thing to touch the connection.

Signed-off-by: Sage Weil <sage@redhat.com>